Move DllInjection in to its own interfaces (#79)
* Move DllInjection in to its own interfaces

* Kill process if injection fails

* [GHA] Format Source

---------
