+++ UnchainedLauncher.Core/Services/Processes/Chivalry/IUnchainedChivalry2Launcher.cs
@@ -22,15 +22,15 @@ namespace UnchainedLauncher.Core.Services.Processes.Chivalry {
 
     public abstract record UnchainedLaunchFailure(string Message, int Code, Option<Error> Underlying) : Expected(Message, Code, Underlying) {
 
-        public record InjectionFailedError(Option<IEnumerable<string>> DllPaths, Error Cause) : UnchainedLaunchFailure($"Failed to inject DLLs '{string.Join("', '", DllPaths)}'", 0, Some(Cause));
+        public record InjectionFailedError() : UnchainedLaunchFailure("Failed to inject process with unchained code.", 0, None);
         public record LaunchFailedError(LaunchFailed Failure) : UnchainedLaunchFailure(Failure.Message, Failure.Code, Failure.Underlying);
         public record DependencyDownloadFailedError(IEnumerable<string> DependencyNames)
             : UnchainedLaunchFailure($"Failed to download dependencies '{string.Join("', '", DependencyNames)}'", 0, None);
         public record LaunchCancelledError() : UnchainedLaunchFailure($"Launch Cancelled", 0, None);
 
 
-        public static UnchainedLaunchFailure InjectionFailed(Option<IEnumerable<string>> dllPaths, Error cause) =>
-            new InjectionFailedError(dllPaths, cause);
+        public static UnchainedLaunchFailure InjectionFailed() =>
+            new InjectionFailedError();
         public static UnchainedLaunchFailure LaunchFailed(LaunchFailed failure) => new LaunchFailedError(failure);
         public static UnchainedLaunchFailure DependencyDownloadFailed(IEnumerable<string> dependencyNames) =>
             new DependencyDownloadFailedError(dependencyNames);
