+++ UnchainedLauncher.GUI/App.xaml.cs
@@ -12,6 +12,7 @@ using UnchainedLauncher.Core.Services.Installer;
 using UnchainedLauncher.Core.Services.Mods;
 using UnchainedLauncher.Core.Services.Mods.Registry;
 using UnchainedLauncher.Core.Services.Mods.Registry.Downloader;
+using UnchainedLauncher.Core.Services.Processes;
 using UnchainedLauncher.Core.Services.Processes.Chivalry;
 using UnchainedLauncher.Core.Utilities;
 using UnchainedLauncher.GUI.Services;
@@ -128,12 +129,8 @@ namespace UnchainedLauncher.GUI {
                 new FileInfoVersionExtractor(),
                 userDialogueSpawner,
                 Directory.GetCurrentDirectory(),
-                () => {
-                    var dllPath = Path.Combine(Directory.GetCurrentDirectory(), FilePaths.PluginDir);
-                    if (!Directory.Exists(dllPath))
-                        Directory.CreateDirectory(dllPath);
-                    return Directory.EnumerateFiles(dllPath);
-                });
+                new DllInjector(Path.Combine(Directory.GetCurrentDirectory(), FilePaths.PluginDir))
+            );
 
             var serversViewModel = new ServersViewModel(settingsViewModel, null);
             var launcherViewModel = new LauncherViewModel(settingsViewModel, modManager, vanillaLauncher, clientsideModdedLauncher, unchainedLauncher, pluginReleaseLocator, new FileInfoVersionExtractor(), userDialogueSpawner);
